Implement video assignment to module and enhance video management features
- Added `assignVideoToModule` method in `HlsVideoController` to handle video assignments to specific modules.
- Updated `deleteVideo` method in `VideoService` to detach videos from models before deletion.
- Enhanced existing video management UI with a new panel for selecting and filtering existing videos.
- Improved code formatting and organization across multiple files for better readability and consistency.

+++ src/Services/VideoService.php
@@ -59,9 +59,19 @@ class VideoService
         return HlsVideo::find($id);
     }
 
-    static function deleteVideo($id)
+    static function deleteVideo($request, $id)
     {
-        return HlsVideo::find($id)->delete();
+        $video = HlsVideo::find($id);
+        $model = $request->modelType ? $request->modelType::find($request->modelId) : null;
+
+        if ($model) {
+            $model->hlsVideos()->detach($id);
+        }
+
+        if (! $video->HlsVideoables()->count() && ! $video->parentFolders()->count()) {
+
+            return HlsVideo::find($id)->delete();
+        }
     }
 
     public function receiveVideo($request, $model = null, $folderId = null)
@@ -118,7 +128,7 @@ class VideoService
             'original_file_name' => $originalFileName ?? $file->getClientOriginalName()
         ]);
 
-        $folder = $folderId
+        $folder = $folderId && $folderId != 'null'
             ? HlsFolder::find($folderId)
             : HlsFolder::whereNull('parent_id')->first();
 
@@ -134,7 +144,7 @@ class VideoService
 
         return $video;
     }
-  
+
     public function receiveFromServer($request, $videoId)
     {
         $video = HlsVideo::findOrFail($videoId);
